logistics_backend/requirements.txt
Django==5.1
djangorestframework==3.16
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
Pillow==12.0.0
python-decouple==3.8